*** empty log message ***
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@3629 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/Zone.m
@@ -379,8 +379,9 @@ PHASE(Using)
 
   [super describe: outputCharStream];
   sprintf (buffer, "> number of objects in population: %u\n"
-           "> total size of objects in population: %u\n",
-           [population getCount], populationTotal);
+           "> total size of objects in population: %lu\n",
+           [population getCount],
+           (unsigned long) populationTotal);
   [outputCharStream catC: buffer];
   
   if (_obj_debug)
